Added check for zero_crossings in case point occurs at end of timeseries causing an Index out of range exception
New function detect_4_by_std, which check if heave displacement values exceed 4
times the standard deviation of a specified set of waves ( currently 2304
equating to half an hour of messages ). If a displacement exceeds this it is
thought to be being acted upon by a breaking wave or other phenomena which will
give unrealistic displacement values.